Remove defaults and pass the value as string
As discussed in this review comment:
https://github.com/CartoDB/grainstore/pull/145/files#r179064748

it is not good to spread defaults around projects. Also, passing the
value as string instead of bool allows for passing it explicitly all
the way down to mapnik, instead of having carto renderer swallow it.

The good thing is that the mapping between config params and the
actual mapnik configuration xml is restricted to the mml builder.